Expose get_node() in the public API
