+++ changes/xxxx.doc.rst
@@ -1,0 +1,1 @@
+Added migration guide for ``zarr.storage.contains_array()`` and ``zarr.storage.contains_group()``.
