+++ changes/xxxx.feature.rst
@@ -1,0 +1,1 @@
+Added `zarr.storage.get_node` to get a node within a store.
