+++ src/zarr/storage/__init__.py
@@ -3,7 +3,7 @@ import warnings
 from types import ModuleType
 from typing import Any
 
-from zarr.storage._common import StoreLike, StorePath
+from zarr.storage._common import StoreLike, StorePath, get_node
 from zarr.storage._fsspec import FsspecStore
 from zarr.storage._local import LocalStore
 from zarr.storage._logging import LoggingStore
@@ -23,6 +23,7 @@ __all__ = [
     "StorePath",
     "WrapperStore",
     "ZipStore",
+    "get_node",
 ]
